+++ tests/test_schema.py
@@ -1,4 +1,5 @@
 import unittest
+import warnings
 import copy
 from pydantic import parse_obj_as, ValidationError
 
@@ -141,6 +142,61 @@ class TestSchema(unittest.TestCase):
             },
         }
 
+        # Non-blended electrodes
+        self.base_non_blended = {
+            "Header": {
+                "BPX": 1.0,
+                "Model": "SPM",
+            },
+            "Parameterisation": {
+                "Cell": {
+                    "Ambient temperature [K]": 299.0,
+                    "Initial temperature [K]": 299.0,
+                    "Reference temperature [K]": 299.0,
+                    "Electrode area [m2]": 2.0,
+                    "External surface area [m2]": 2.2,
+                    "Volume [m3]": 1.0,
+                    "Number of electrode pairs connected in parallel to make a cell": 1,
+                    "Nominal cell capacity [A.h]": 5.0,
+                    "Lower voltage cut-off [V]": 2.0,
+                    "Upper voltage cut-off [V]": 4.0,
+                },
+                "Negative electrode": {
+                    "Particle radius [m]": 5.86e-6,
+                    "Thickness [m]": 85.2e-6,
+                    "Diffusivity [m2.s-1]": 3.3e-14,
+                    "OCP [V]": (
+                        "9.47057878e-01 * exp(-1.59418743e+02  * x) - 3.50928033e+04 + "
+                        "1.64230269e-01 * tanh(-4.55509094e+01 * (x - 3.24116012e-02 )) + "
+                        "3.69968491e-02 * tanh(-1.96718868e+01 * (x - 1.68334476e-01)) + "
+                        "1.91517003e+04 * tanh(3.19648312e+00 * (x - 1.85139824e+00)) + "
+                        "5.42448511e+04 * tanh(-3.19009848e+00 * (x - 2.01660395e+00))"
+                    ),
+                    "Surface area per unit volume [m-1]": 383959,
+                    "Reaction rate constant [mol.m-2.s-1]": 1e-10,
+                    "Maximum concentration [mol.m-3]": 33133,
+                    "Minimum stoichiometry": 0.005504,
+                    "Maximum stoichiometry": 0.75668,
+                },
+                "Positive electrode": {
+                    "Particle radius [m]": 5.22e-6,
+                    "Thickness [m]": 75.6e-6,
+                    "Diffusivity [m2.s-1]": 4.0e-15,
+                    "OCP [V]": (
+                        "-3.04420906 * x + 10.04892207 - "
+                        "0.65637536 * tanh(-4.02134095 * (x - 0.80063948)) + "
+                        "4.24678547 * tanh(12.17805062 * (x - 7.57659337)) - "
+                        "0.3757068 * tanh(59.33067782 * (x - 0.99784492))"
+                    ),
+                    "Surface area per unit volume [m-1]": 382184,
+                    "Reaction rate constant [mol.m-2.s-1]": 1e-10,
+                    "Maximum concentration [mol.m-3]": 63104.0,
+                    "Minimum stoichiometry": 0.42424,
+                    "Maximum stoichiometry": 0.96210,
+                },
+            },
+        }
+
     def test_simple(self):
         test = copy.copy(self.base)
         parse_obj_as(BPX, test)
@@ -257,6 +313,39 @@ class TestSchema(unittest.TestCase):
                 "Temperature [K]": [298.15, 298.15],
             },
         }
+
+    def test_check_sto_limits_validator(self):
+        warnings.filterwarnings("error")  # Treat warnings as errors
+        test = copy.copy(self.base_non_blended)
+        test["Parameterisation"]["Cell"]["Upper voltage cut-off [V]"] = 4.3
+        test["Parameterisation"]["Cell"]["Lower voltage cut-off [V]"] = 2.5
+        parse_obj_as(BPX, test)
+
+    def test_check_sto_limits_validator_high_voltage(self):
+        test = copy.copy(self.base_non_blended)
+        test["Parameterisation"]["Cell"]["Upper voltage cut-off [V]"] = 4.0
+        with self.assertWarns(UserWarning):
+            parse_obj_as(BPX, test)
+
+    def test_check_sto_limits_validator_high_voltage_tolerance(self):
+        warnings.filterwarnings("error")  # Treat warnings as errors
+        test = copy.copy(self.base_non_blended)
+        test["Parameterisation"]["Cell"]["Upper voltage cut-off [V]"] = 4.0
+        BPX.settings.tolerances["Voltage [V]"] = 0.25
+        parse_obj_as(BPX, test)
+
+    def test_check_sto_limits_validator_low_voltage(self):
+        test = copy.copy(self.base_non_blended)
+        test["Parameterisation"]["Cell"]["Lower voltage cut-off [V]"] = 3.0
+        with self.assertWarns(UserWarning):
+            parse_obj_as(BPX, test)
+
+    def test_check_sto_limits_validator_low_voltage_tolerance(self):
+        warnings.filterwarnings("error")  # Treat warnings as errors
+        test = copy.copy(self.base_non_blended)
+        test["Parameterisation"]["Cell"]["Lower voltage cut-off [V]"] = 3.0
+        BPX.settings.tolerances["Voltage [V]"] = 0.35
+        parse_obj_as(BPX, test)
 
     def test_user_defined(self):
         test = copy.copy(self.base)